+++ HISTORY.rst
@@ -25,6 +25,17 @@ History
   Two additional configuration parameters, ``SmaltIndex_ExtraParams`` and ``SmaltAlign_ExtraParams`` 
   can be configured with any Smalt command line options.  See :ref:`tool-selection-label`.  The
   default aligner is still bowtie2.
+* Split the create_snp_matrix.py script into two pieces.  The new script, call_consensus.py, is a redesigned
+  consensus caller which can be run in parallel to call snps for multiple samples concurrently.  The
+  create_snp_matrix.py script smiply merges the consensus calls for all samples into a multi-fasta file.
+* The new consensus caller has the following adjustable parameters.  
+  See the :ref:`cmd-ref-call-consensus` command reference.
+
+  * ``minBaseQual`` : Mimimum base quality score to count a read.
+  * ``minConsFreq`` : Minimum consensus frequency.
+  * ``minConsStrdDpth`` : Minimum consensus-supporting strand depth.
+  * ``minConsStrdBias``: Strand bias.
+
 
 0.3.4 (2015-06-25)
 ~~~~~~~~~~~~~~~~~~